requirements.txt
asgiref==3.4.1; python_version >= "3.6"
beautifulsoup4==4.10.0; python_full_version > "3.0.0"
bs4==0.0.1
certifi==2021.10.8; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0"
charset-normalizer==2.0.9; python_full_version >= "3.6.0" and python_version >= "3"
datetime==4.3
django-cors-headers==3.10.1; python_version >= "3.6"
django-environ==0.8.1; python_version >= "3.4" and python_version < "4"
django-nose==1.4.7
django==4.0; python_version >= "3.8"
djangorestframework-simplejwt==5.0.0; python_version >= "3.7"
djangorestframework==3.13.0; python_version >= "3.6"
gunicorn==20.1.0; python_version >= "3.5"
idna==3.3; python_version >= "3.5" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.5"
nose==1.3.7
numpy==1.21.4
pandas==1.3.5; python_full_version >= "3.7.1"
psycopg2-binary==2.9.2; python_version >= "3.6"
pyjwt==2.3.0; python_version >= "3.6"
python-dateutil==2.8.2; python_full_version >= "3.7.1"
pytz==2021.3
requests==2.26.0; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.6.0")
six==1.16.0; python_full_version >= "3.7.1"
soupsieve==2.3.1; python_version >= "3.6" and python_full_version > "3.0.0"
sqlparse==0.4.2; python_version >= "3.5"
tzdata==2021.5; sys_platform == "win32" and python_version >= "3.8"
urllib3==1.26.7; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version < "4"
whitenoise==5.3.0; python_version >= "3.5" and python_version < "4"
zope.interface==5.4.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
